+++ spec/dummy/config/initializers/jwt.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+JWT.configuration.jwk.kid_generator_type = :rfc7638_thumbprint